TFKUBE-228: Refactored the bamboo e2e tests (#106)
* TFKUBE-228: Refactored the bamboo e2e tests

* TFKUBE-228: Changed the code to fix the lint errors
